+++ tests/Handler/BindingCommandHandlerTest.php
@@ -13,7 +13,7 @@ namespace Puli\Cli\Tests\Handler;
 
 use PHPUnit_Framework_Assert;
 use PHPUnit_Framework_MockObject_MockObject;
-use Puli\Cli\Handler\BindCommandHandler;
+use Puli\Cli\Handler\BindingCommandHandler;
 use Puli\Manager\Api\Discovery\BindingDescriptor;
 use Puli\Manager\Api\Discovery\BindingState;
 use Puli\Manager\Api\Discovery\DiscoveryManager;
@@ -32,7 +32,7 @@ use Webmozart\Expression\Expression;
  * @since  1.0
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-class BindCommandHandlerTest extends AbstractCommandHandlerTest
+class BindingCommandHandlerTest extends AbstractCommandHandlerTest
 {
     const BINDING_UUID1 = 'bb5a07d6-e979-4c35-9883-4d8c1165e3d5';
     const BINDING_UUID2 = 'cc9f2259-8587-4d4e-9dda-f0ff87b4e871';
@@ -63,7 +63,7 @@ class BindCommandHandlerTest extends AbstractCommandHandlerTest
     /**
      * @var Command
      */
-    private static $deleteCommand;
+    private static $removeCommand;
 
     /**
      * @var Command
@@ -86,7 +86,7 @@ class BindCommandHandlerTest extends AbstractCommandHandlerTest
     private $packages;
 
     /**
-     * @var BindCommandHandler
+     * @var BindingCommandHandler
      */
     private $handler;
 
@@ -94,11 +94,11 @@ class BindCommandHandlerTest extends AbstractCommandHandlerTest
     {
         parent::setUpBeforeClass();
 
-        self::$listCommand = self::$application->getCommand('bind')->getSubCommand('list');
-        self::$saveCommand = self::$application->getCommand('bind')->getSubCommand('save');
-        self::$deleteCommand = self::$application->getCommand('bind')->getSubCommand('delete');
-        self::$enableCommand = self::$application->getCommand('bind')->getSubCommand('enable');
-        self::$disableCommand = self::$application->getCommand('bind')->getSubCommand('disable');
+        self::$listCommand = self::$application->getCommand('binding')->getSubCommand('list');
+        self::$saveCommand = self::$application->getCommand('binding')->getSubCommand('add');
+        self::$removeCommand = self::$application->getCommand('binding')->getSubCommand('remove');
+        self::$enableCommand = self::$application->getCommand('binding')->getSubCommand('enable');
+        self::$disableCommand = self::$application->getCommand('binding')->getSubCommand('disable');
     }
 
     protected function setUp()
@@ -111,7 +111,7 @@ class BindCommandHandlerTest extends AbstractCommandHandlerTest
             new Package(new PackageFile('vendor/package1'), '/package1'),
             new Package(new PackageFile('vendor/package2'), '/package2'),
         ));
-        $this->handler = new BindCommandHandler($this->discoveryManager, $this->packages);
+        $this->handler = new BindingCommandHandler($this->discoveryManager, $this->packages);
     }
 
     public function testListAllBindings()
@@ -759,7 +759,7 @@ EOF;
 
     public function testRemoveBinding()
     {
-        $args = self::$deleteCommand->parseArgs(new StringArgs('ab12'));
+        $args = self::$removeCommand->parseArgs(new StringArgs('ab12'));
         $descriptor = new BindingDescriptor('/path', 'my/type', array(), 'glob');
 
         $this->discoveryManager->expects($this->once())
@@ -773,7 +773,7 @@ EOF;
             ->method('removeBinding')
             ->with($descriptor->getUuid());
 
-        $statusCode = $this->handler->handleDelete($args, $this->io);
+        $statusCode = $this->handler->handleRemove($args, $this->io);
 
         $this->assertSame(0, $statusCode);
         $this->assertEmpty($this->io->fetchOutput());
@@ -786,7 +786,7 @@ EOF;
      */
     public function testRemoveBindingFailsIfAmbiguous()
     {
-        $args = self::$deleteCommand->parseArgs(new StringArgs('ab12'));
+        $args = self::$removeCommand->parseArgs(new StringArgs('ab12'));
 
         $this->discoveryManager->expects($this->once())
             ->method('findBindings')
@@ -801,7 +801,7 @@ EOF;
         $this->discoveryManager->expects($this->never())
             ->method('removeBinding');
 
-        $this->handler->handleDelete($args, $this->io);
+        $this->handler->handleRemove($args, $this->io);
     }
 
     /**
@@ -810,7 +810,7 @@ EOF;
      */
     public function testRemoveBindingFailsIfNotFound()
     {
-        $args = self::$deleteCommand->parseArgs(new StringArgs('ab12'));
+        $args = self::$removeCommand->parseArgs(new StringArgs('ab12'));
 
         $this->discoveryManager->expects($this->at(0))
             ->method('findBindings')
@@ -829,7 +829,7 @@ EOF;
         $this->discoveryManager->expects($this->never())
             ->method('removeBinding');
 
-        $this->handler->handleDelete($args, $this->io);
+        $this->handler->handleRemove($args, $this->io);
     }
 
     /**
@@ -838,7 +838,7 @@ EOF;
      */
     public function testRemoveBindingFailsIfNoRootBinding()
     {
-        $args = self::$deleteCommand->parseArgs(new StringArgs('ab12'));
+        $args = self::$removeCommand->parseArgs(new StringArgs('ab12'));
 
         $this->discoveryManager->expects($this->at(0))
             ->method('findBindings')
@@ -857,7 +857,7 @@ EOF;
         $this->discoveryManager->expects($this->never())
             ->method('removeBinding');
 
-        $this->handler->handleDelete($args, $this->io);
+        $this->handler->handleRemove($args, $this->io);
     }
 
     public function testEnableBindings()