+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog
  * renamed `--type` option to `--class` for `puli find`
  * renamed `--bound-to` option to `--type` for `puli find`
  * removed the `--package` option of `puli bind --enable|--disable`
+ * renamed `puli bind` to `puli binding add|list|remove|enable|disable`
 
 * 1.0.0-beta3 (2015-03-19)